+++ Forge/src/app/models/portfolio.ts
@@ -1,29 +1,13 @@
-export interface Portfolio {
-    userId?: number,
-    email?: String,
-    password?: String,
-    firstName?: String,
-    lastName?: String,
-    portfolios?: [
-        {
-            id?: number,
-            belongsTo?: String,
-            status?: String,
-            aboutMe?: {
-                id?: number,
-                description?: String,
-                aboutMeItems?: [
-                    {
-                        id?: number,
-                        content?: String
-                    }
-                ]
-            },
-            projects?: [],
-            education?: [],
-            skillMatrix?: [],
-            industryEquivalency?: []
-        }
-    ],
-    admin?: boolean
+import { User } from "../models/user";
+
+export class Portfolio {
+    id: number;
+    belongsTo: string;
+    status: string;
+    aboutMe: Object;
+    industryEquivalency: Object;
+    myUser?: User;
+    projects: Object;
+    education: Object;
+    skillMatrix: Object;
 }
